+++ src/scripts/InstanceScripts/Wotlk/IcecrownCitadel/LordMarrowgar.cpp
@@ -3,7 +3,7 @@ Copyright (c) 2014-2022 AscEmu Team <http://www.ascemu.org>
 This file is released under the MIT license. See README-MIT for more information.
 */
 
-#include "LordMarrowgar.h"
+#include "LordMarrowgar.hpp"
 #include "Management/Faction.h"
 #include "Objects/Units/Creatures/Summons/Summon.h"
 #include <Management/ObjectMgr.h>
