+++ src/scripts/InstanceScripts/Wotlk/IcecrownCitadel/LordMarrowgar.hpp
@@ -5,7 +5,7 @@ This file is released under the MIT license. See README-MIT for more information
 
 #pragma once
 
-#include "Raid_IceCrownCitadel.h"
+#include "Raid_IceCrownCitadel.hpp"
 
 //////////////////////////////////////////////////////////////////////////////////////////
 /// Boss: Lord Marrowgar
@@ -14,7 +14,6 @@ class LordMarrowgarAI : public CreatureAIScript
 public:
     static CreatureAIScript* Create(Creature* pCreature);
     explicit LordMarrowgarAI(Creature* pCreature);
-
 
     void IntroStart();
